perf(server): share tiny client batch retention paths
